create separate filesList object
