+++ tests/acceptance/pageObjects/filesPage.js
@@ -8,12 +8,7 @@ module.exports = {
      * @param {string} folder
      */
     navigateToFolder: function (folder) {
-      this.waitForFileVisible(folder)
-      this.useXpath()
-        .moveToElement(this.getFileRowSelectorByFileName(folder), 0, 0)
-        .click(this.getFileLinkSelectorByFileName(folder))
-        .useCss()
-      return this.waitForElementNotPresent('@filesListProgressBar')
+      return this.page.FilesPageElement.filesList().navigateToFolder(folder)
         .waitForElementVisible('@breadcrumb')
         .assert.containsText('@breadcrumb', folder)
     },
@@ -39,63 +34,6 @@ module.exports = {
     },
     /**
      *
-     * @param {string} fileName
-     */
-    deleteFile: function (fileName) {
-      const deleteBtnSelector = this.getFileRowSelectorByFileName(fileName) +
-                                this.elements['deleteButtonInFileRow'].selector
-      return this.initAjaxCounters()
-        .waitForFileVisible(fileName)
-        .useXpath()
-        .moveToElement(this.getFileRowSelectorByFileName(fileName), 0, 0)
-        .click(deleteBtnSelector)
-        .waitForElementVisible('@deleteFileConfirmationBtn')
-        .waitForAnimationToFinish()
-        .click('@deleteFileConfirmationBtn')
-        .waitForElementNotVisible('@deleteFileConfirmationDialog')
-        .waitForOutstandingAjaxCalls()
-        .useCss()
-    },
-    /**
-     *
-     * @param {string} fileName
-     */
-    openSharingDialog: function (fileName) {
-      const shareBtnSelector = this.getFileRowSelectorByFileName(fileName) +
-        this.elements['shareButtonInFileRow'].selector
-
-      this.waitForFileVisible(fileName)
-        .useXpath()
-        .moveToElement(this.getFileRowSelectorByFileName(fileName), 0, 0)
-        .click(shareBtnSelector)
-        .waitForElementVisible('@sharingSideBar')
-        .useCss()
-      return this.api.page.FilesPageElement.sharingDialog()
-    },
-    /**
-     *
-     * @param {string} fromName
-     * @param {string} toName
-     */
-    renameFile: function (fromName, toName) {
-      const renameBtnSelector = this.getFileRowSelectorByFileName(fromName) +
-        this.elements['renameButtonInFileRow'].selector
-      return this.initAjaxCounters()
-        .waitForFileVisible(fromName)
-        .useXpath()
-        .moveToElement(this.getFileRowSelectorByFileName(fromName), 0, 0)
-        .click(renameBtnSelector)
-        .waitForElementVisible('@renameFileConfirmationBtn')
-        .waitForAnimationToFinish()
-        .clearValue('@renameFileInputField')
-        .setValue('@renameFileInputField', toName)
-        .click('@renameFileConfirmationBtn')
-        .waitForElementNotVisible('@renameFileConfirmationDialog')
-        .waitForOutstandingAjaxCalls()
-        .useCss()
-    },
-    /**
-     *
      * @param {string} localFileName
      */
     uploadFile: function (localFileName) {
@@ -116,68 +54,12 @@ module.exports = {
         .waitForElementNotVisible('@fileUploadProgress')
         .click('@newFileMenuButton')
     },
-    /**
-     *
-     * @param {string} fileName
-     */
-    waitForFileVisible: function (fileName) {
-      const rowSelector = this.getFileRowSelectorByFileName(fileName)
-      const linkSelector = this.getFileLinkSelectorByFileName(fileName)
-      this
-        .useXpath()
-        .waitForElementVisible(rowSelector)
-        .expect.element(linkSelector).text.to.equal(fileName)
-      return this.useCss()
-    },
-    /**
-     *
-     * @param {string} fileName
-     *
-     * @returns {string}
-     */
-    getFileRowSelectorByFileName: function (fileName) {
-      var element = this.elements['fileRowByName']
-      var util = require('util')
-      if (fileName.indexOf('"') > -1) {
-        element.selector = element.selector.replace(/"/g, "'")
-      }
-      return util.format(element.selector, fileName)
-    },
-    /**
-     *
-     * @param {string} fileName
-     */
-    getFileLinkSelectorByFileName: function (fileName) {
-      return this
-        .getFileRowSelectorByFileName(fileName) +
-          this.elements['fileLinkInFileRow'].selector
-    },
     showHiddenFiles: function () {
       return this
         .waitForElementVisible('@filterListButton')
         .click('@filterListButton')
         .waitForElementVisible('@hiddenFilesLabel')
         .click('@hiddenFilesCheckbox')
-    },
-    /**
-     *
-     * @param {function} callback
-     */
-    allFileRows: function (callback) {
-      this.api.elements('css selector', this.elements['fileRows'], function (result) {
-        callback(result)
-      })
-    },
-    /**
-     *
-     * @param {string} element
-     */
-    assertElementNotListed: function (element) {
-      return this
-        .waitForElementVisible('@filesTable')
-        .useXpath()
-        .waitForElementNotPresent('@loadingIndicator')
-        .waitForElementNotPresent(this.getFileRowSelectorByFileName(element))
     },
     /**
      *
@@ -201,61 +83,9 @@ module.exports = {
         .waitForElementVisible(labelSelector)
         .toggleCheckbox(enableOrDisable, checkboxId)
         .click('@filterListButton')
-    },
-    /**
-     *
-     * @param {string} path
-     */
-    markAsFavorite: function (path) {
-      const favoriteIconButton = this.getFileRowSelectorByFileName(path) +
-        this.elements['notMarkedFavoriteInFileRow'].selector
-
-      return this.initAjaxCounters()
-        .waitForFileVisible(path)
-        .useXpath()
-        .click(favoriteIconButton)
-        .waitForOutstandingAjaxCalls()
-        .useCss()
-    },
-    /**
-     *
-     */
-    unmarkFavorite: function (path) {
-      const unFavoriteBtn = this.getFileRowSelectorByFileName(path) +
-        this.elements['markedFavoriteInFileRow'].selector
-
-      return this.initAjaxCounters()
-        .waitForFileVisible(path)
-        .useXpath()
-        .click(unFavoriteBtn)
-        .waitForOutstandingAjaxCalls()
-        .useCss()
-    },
-    /**
-     *
-     * @param {string} path
-     * @param {function(boolean)} callback - if the file is marked as favorite
-     */
-    isMarkedFavorite: function (path, callback) {
-      const markedFavoriteIcon = this.getFileRowSelectorByFileName(path) +
-        this.elements['markedFavoriteInFileRow'].selector
-
-      return this
-        .waitForFileVisible(path)
-        .useXpath()
-        .isVisible(markedFavoriteIcon, (result) => {
-          callback(result.value)
-        })
-        .useCss()
     }
   },
   elements: {
-    filesTable: {
-      selector: '#files-list'
-    },
-    fileRows: {
-      selector: 'tr.file-row'
-    },
     newFileMenuButton: {
       selector: '#new-file-menu-btn'
     },
@@ -271,45 +101,9 @@ module.exports = {
     newFolderOkButton: {
       selector: '#new-folder-ok'
     },
-    filesListProgressBar: {
-      selector: '#files-list-progress'
-    },
+
     breadcrumb: {
       selector: '#files-breadcrumb li:nth-of-type(1)'
-    },
-    fileRowByName: {
-      selector: '//a[contains(@class, "file-row-name")][@filename="%s"]/../..'
-    },
-    fileLinkInFileRow: {
-      selector: '//a[contains(@class, "file-row-name")]'
-    },
-    deleteFileConfirmationDialog: {
-      selector: '#delete-file-confirmation-dialog'
-    },
-    deleteButtonInFileRow: {
-      selector: '//button[@aria-label="Delete"]',
-      locateStrategy: 'xpath'
-    },
-    shareButtonInFileRow: {
-      selector: '//button[@aria-label="Share"]',
-      locateStrategy: 'xpath'
-    },
-    deleteFileConfirmationBtn: {
-      selector: '#oc-dialog-delete-confirm'
-    },
-    renameFileConfirmationDialog: {
-      selector: '#change-file-dialog'
-    },
-    renameButtonInFileRow: {
-      selector: '//button[@aria-label="Edit"]',
-      locateStrategy: 'xpath'
-    },
-    renameFileInputField: {
-      selector: '//div[@id="change-file-dialog"]//input',
-      locateStrategy: 'xpath'
-    },
-    renameFileConfirmationBtn: {
-      selector: '#oc-dialog-rename-confirm'
     },
     filterListButton: {
       selector: '#oc-filter-list-btn'
@@ -332,10 +126,6 @@ module.exports = {
     filterFileCheckbox: {
       selector: '#oc-filter-file-checkbox'
     },
-    loadingIndicator: {
-      selector: '//*[contains(@class, "oc-loader")]',
-      locateStrategy: 'xpath'
-    },
     createFolderLoadingIndicator: {
       selector: '//div[@id="new-folder-dialog"]//div[@class="oc-loader"]',
       locateStrategy: 'xpath'
@@ -349,17 +139,6 @@ module.exports = {
     },
     fileUploadProgress: {
       selector: '#oc-progress-pie'
-    },
-    notMarkedFavoriteInFileRow: {
-      selector: '//span[contains(@class, "oc-star-dimm")]',
-      locateStrategy: 'xpath'
-    },
-    markedFavoriteInFileRow: {
-      selector: '//span[contains(@class, "oc-star-shining")]',
-      locateStrategy: 'xpath'
-    },
-    sharingSideBar: {
-      selector: '#oc-files-sharing-sidebar'
     }
   }
 }
