+++ tests/acceptance/stepDefinitions/loginContext.js
@@ -28,7 +28,7 @@ const loginAsUser = function (userId) {
 
   // Then the files table should be displayed
   return client
-    .page.filesPage()
+    .page.FilesPageElement.filesList()
     .waitForElementVisible('@filesTable')
     .then(() => {
       client.globals.currentUserName = userId
@@ -72,8 +72,7 @@ When('the user authorizes access to phoenix',
 
 Then('the files table should not be empty',
   () => {
-    const filesPage = client.page.filesPage()
-    return filesPage
+    return client.page.FilesPageElement.filesList()
     // even the loading indicator is gone the table might not be rendered yet
       .waitForElementVisible('@fileRows')
   })
